databases/requirements.txt
soccerdata==1.8.7
pandas>=2.0
psycopg2-binary>=2.9
SQLAlchemy>=2.0
python-dateutil
python-dotenv
